favorite workouts now stored in a Drawer to be shown when user presses the Favorites button
